requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
feedparser==6.10.0
aiohttp==3.8.4
beautifulsoup4==4.12.2